Simplify tests for sorting algorithms
