+++ sorting_algorithms/__init__.py
@@ -1,0 +1,3 @@
+from .selection_sort import SelectionSort
+from .bubble_sort import BubbleSort
+from .quicksort import Quicksort
